backend/requirements.txt
django
djangorestframework
django-cors-headers
moviepy 
requests
ffmpeg
vosk
librosa